+++ controllers/index.js
@@ -1,0 +1,5 @@
+const contactControllers = require("./contactControllers")
+
+module.exports = {
+    contactControllers
+}
