+++ models/contacts.json
@@ -1,1 +1,1 @@
-[{"id":"AeHIrLTr6JkxGE6SN-0Rw","name":"Allen Raymond","email":"nulla.ante@vestibul.co.uk","phone":"(992) 914-3792"},{"id":"qdggE76Jtbfd9eWJHrssH","name":"Chaim Lewis","email":"dui.in@egetlacus.ca","phone":"(294) 840-6685"},{"id":"drsAJ4SHPYqZeG-83QTVW","name":"Kennedy Lane","email":"mattis.Cras@nonenimMauris.net","phone":"(542) 451-7038"},{"id":"vza2RIzNGIwutCVCs4mCL","name":"Wylie Pope","email":"est@utquamvel.net","phone":"(692) 802-2949"},{"id":"05olLMgyVQdWRwgKfg5J6","name":"Cyrus Jackson","email":"nibh@semsempererat.com","phone":"(501) 472-5218"},{"id":"1DEXoP8AuCGYc1YgoQ6hw","name":"Abbot Franks","email":"scelerisque@magnis.org","phone":"(186) 568-3720"},{"id":"Z5sbDlS7pCzNsnAHLtDJd","name":"Reuben Henry","email":"pharetra.ut@dictum.co.uk","phone":"(715) 598-5792"},{"id":"C9sjBfCo4UJCWjzBnOtxl","name":"Simon Morton","email":"dui.Fusce.diam@Donec.com","phone":"(233) 738-2360"},{"id":"e6ywwRe4jcqxXfCZOj_1e","name":"Thomas Lucas","email":"nec@Nulla.com","phone":"(704) 398-7993"},{"id":"rsKkOQUi80UsgVPCcLZZW","name":"Alec Howard","email":"Donec.elementum@scelerisquescelerisquedui.net","phone":"(748) 206-2688"},{"name":"Max Dimekhin","email":"hui@g.com","phone":"45-45-34","id":"f5705da4-a99a-48b6-84f4-31812372aee8"},{"name":"Not that mad Max","email":"notThatMadMax@gmail.com","phone":"+380677755688","id":"0ba7ca33-bfc5-4bf0-8524-111c769cfecc"},{"name":"Not that mad Max","email":"notThatMadMax@gmail.com","phone":"+380677755688","id":"ce6076ab-e658-43c3-a67f-4bdc2caa518a"}]
+[{"id":"AeHIrLTr6JkxGE6SN-0Rw","name":"Allen Raymond","email":"nulla.ante@vestibul.co.uk","phone":"(992) 914-3792"},{"id":"qdggE76Jtbfd9eWJHrssH","name":"Chaim Lewis","email":"dui.in@egetlacus.ca","phone":"(294) 840-6685"},{"id":"drsAJ4SHPYqZeG-83QTVW","name":"Kennedy Lane","email":"mattis.Cras@nonenimMauris.net","phone":"(542) 451-7038"},{"id":"vza2RIzNGIwutCVCs4mCL","name":"Wylie Pope","email":"est@utquamvel.net","phone":"(692) 802-2949"},{"id":"05olLMgyVQdWRwgKfg5J6","name":"Cyrus Jackson","email":"nibh@semsempererat.com","phone":"(501) 472-5218"},{"id":"1DEXoP8AuCGYc1YgoQ6hw","name":"Abbot Franks","email":"scelerisque@magnis.org","phone":"(186) 568-3720"},{"id":"Z5sbDlS7pCzNsnAHLtDJd","name":"Reuben Henry","email":"pharetra.ut@dictum.co.uk","phone":"(715) 598-5792"},{"id":"C9sjBfCo4UJCWjzBnOtxl","name":"Simon Morton","email":"dui.Fusce.diam@Donec.com","phone":"(233) 738-2360"},{"id":"e6ywwRe4jcqxXfCZOj_1e","name":"Thomas Lucas","email":"nec@Nulla.com","phone":"(704) 398-7993"},{"id":"rsKkOQUi80UsgVPCcLZZW","name":"Alec Howard","email":"Donec.elementum@scelerisquescelerisquedui.net","phone":"(748) 206-2688"},{"name":"Max Dimekhin","email":"hui@g.com","phone":"45-45-34","id":"f5705da4-a99a-48b6-84f4-31812372aee8"},{"name":"Not that mad Max","email":"notThatMadMax@gmail.com","phone":"+380677755688","id":"0ba7ca33-bfc5-4bf0-8524-111c769cfecc"},{"name":"Not that mad Max","email":"notThatMadMax@gmail.com","phone":"+380677755688","id":"ce6076ab-e658-43c3-a67f-4bdc2caa518a"},{"name":"PapaHET","email":"papahet@gmail.com","phone":"25789","id":"45631486-e898-492b-98a4-d46dd89cdfce"},{"name":"Not that mad Max","email":"notThatMadMax@gmail.com","phone":"+380677755688","id":"8e118bc9-592c-4edb-8950-b434b10db11d"}]
