+++ routes/api/index.js
@@ -1,0 +1,5 @@
+const contactRouter = require("./contacts")
+
+module.exports = {
+    contactRouter
+}
